+++ src/cube/permutations.rs
@@ -1,49 +1,49 @@
 use crate::{NUM_CORNERS, NUM_EDGES};
 
 /// The rotation matrix for the front-face corners; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_FRONT_CORNERS: [usize; NUM_CORNERS] = [0, 6, 2, 4, 1, 5, 3, 7];
+pub const PERMUTE_FRONT_CORNERS: [u8; NUM_CORNERS] = [0, 6, 2, 4, 1, 5, 3, 7];
 
 /// The rotation matrix for the front-face edges; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_FRONT_EDGES: [usize; NUM_EDGES] = [0, 1, 2, 3, 4, 9, 8, 7, 5, 6, 10, 11];
+pub const PERMUTE_FRONT_EDGES: [u8; NUM_EDGES] = [0, 1, 2, 3, 4, 9, 8, 7, 5, 6, 10, 11];
 
 /// The rotation matrix for the back-face corner; where old[idx] and new[idx] are swapped.
 ///
 /// NOTE: This rotation is documented incorrectly in http://joren.ralphdesign.nl/projects/rubiks_cube/cube.pdf.
-pub const PERMUTE_BACK_CORNERS: [usize; NUM_CORNERS] = [7, 1, 5, 3, 4, 0, 6, 2];
+pub const PERMUTE_BACK_CORNERS: [u8; NUM_CORNERS] = [7, 1, 5, 3, 4, 0, 6, 2];
 
 /// The rotation matrix for the back-face edges; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_BACK_EDGES: [usize; NUM_EDGES] = [0, 1, 2, 3, 11, 5, 6, 10, 8, 9, 4, 7];
+pub const PERMUTE_BACK_EDGES: [u8; NUM_EDGES] = [0, 1, 2, 3, 11, 5, 6, 10, 8, 9, 4, 7];
 
 /// The rotation matrix for the left-face corners; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_LEFT_CORNERS: [usize; NUM_CORNERS] = [5, 4, 2, 3, 0, 1, 6, 7];
+pub const PERMUTE_LEFT_CORNERS: [u8; NUM_CORNERS] = [5, 4, 2, 3, 0, 1, 6, 7];
 
 /// The rotation matrix for the left-face edges; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_LEFT_EDGES: [usize; NUM_EDGES] = [4, 5, 2, 3, 1, 0, 6, 7, 8, 9, 10, 11];
+pub const PERMUTE_LEFT_EDGES: [u8; NUM_EDGES] = [4, 5, 2, 3, 1, 0, 6, 7, 8, 9, 10, 11];
 
 /// The rotation matrix for the right-face corners; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_RIGHT_CORNERS: [usize; NUM_CORNERS] = [0, 1, 7, 6, 4, 5, 2, 3];
+pub const PERMUTE_RIGHT_CORNERS: [u8; NUM_CORNERS] = [0, 1, 7, 6, 4, 5, 2, 3];
 
 /// The rotation matrix for the right-face edges; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_RIGHT_EDGES: [usize; NUM_EDGES] = [0, 1, 6, 7, 4, 5, 3, 2, 8, 9, 10, 11];
+pub const PERMUTE_RIGHT_EDGES: [u8; NUM_EDGES] = [0, 1, 6, 7, 4, 5, 3, 2, 8, 9, 10, 11];
 
 /// The rotation matrix for the up-face corners; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_UP_CORNERS: [usize; NUM_CORNERS] = [4, 1, 2, 7, 3, 5, 6, 0];
+pub const PERMUTE_UP_CORNERS: [u8; NUM_CORNERS] = [4, 1, 2, 7, 3, 5, 6, 0];
 
 /// The rotation matrix for the up-face edges; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_UP_EDGES: [usize; NUM_EDGES] = [8, 1, 11, 3, 4, 5, 6, 7, 2, 9, 10, 0];
+pub const PERMUTE_UP_EDGES: [u8; NUM_EDGES] = [8, 1, 11, 3, 4, 5, 6, 7, 2, 9, 10, 0];
 
 /// The rotation matrix for the down-face corners; where old[idx] and new[idx] are swapped.
-pub const PERMUTE_DOWN_CORNERS: [usize; NUM_CORNERS] = [0, 5, 6, 3, 4, 2, 1, 7];
+pub const PERMUTE_DOWN_CORNERS: [u8; NUM_CORNERS] = [0, 5, 6, 3, 4, 2, 1, 7];
 
 /// The rotation matrix for the down-face edges where rotation = old + new % orientations.
-pub const PERMUTE_DOWN_EDGES: [usize; NUM_EDGES] = [0, 10, 2, 9, 4, 5, 6, 7, 8, 1, 3, 11];
+pub const PERMUTE_DOWN_EDGES: [u8; NUM_EDGES] = [0, 10, 2, 9, 4, 5, 6, 7, 8, 1, 3, 11];
 
 /// Composes the given single 90 degree turn's permutation with itself `times` times, producing the permutation
 /// equivalent to applying that 90 degree turn `times` times in a row. Used below to derive each face's 180/270
 /// degree tables from its 90 degree one at compile time, rather than looping `Cube::rotate_*` 2-3 times per move
 /// at runtime.
-const fn compose<const N: usize>(rot: [usize; N], times: usize) -> [usize; N] {
-    let mut out = [0usize; N];
+const fn compose<const N: usize>(rot: [u8; N], times: usize) -> [u8; N] {
+    let mut out = [0u8; N];
     let mut i = 0;
 
     while i < N {
@@ -51,11 +51,11 @@ const fn compose<const N: usize>(rot: [usize; N], times: usize) -> [usize; N] {
         let mut t = 0;
 
         while t < times {
-            idx = rot[idx];
+            idx = rot[idx] as usize;
             t += 1;
         }
 
-        out[i] = idx;
+        out[i] = idx as u8;
         i += 1;
     }
 
@@ -63,73 +63,73 @@ const fn compose<const N: usize>(rot: [usize; N], times: usize) -> [usize; N] {
 }
 
 /// The rotation matrix for a 180 degree turn of the front-face corners.
-pub const PERMUTE_FRONT_CORNERS_180: [usize; NUM_CORNERS] = compose(PERMUTE_FRONT_CORNERS, 2);
+pub const PERMUTE_FRONT_CORNERS_180: [u8; NUM_CORNERS] = compose(PERMUTE_FRONT_CORNERS, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the front-face corners.
-pub const PERMUTE_FRONT_CORNERS_270: [usize; NUM_CORNERS] = compose(PERMUTE_FRONT_CORNERS, 3);
+pub const PERMUTE_FRONT_CORNERS_270: [u8; NUM_CORNERS] = compose(PERMUTE_FRONT_CORNERS, 3);
 
 /// The rotation matrix for a 180 degree turn of the front-face edges.
-pub const PERMUTE_FRONT_EDGES_180: [usize; NUM_EDGES] = compose(PERMUTE_FRONT_EDGES, 2);
+pub const PERMUTE_FRONT_EDGES_180: [u8; NUM_EDGES] = compose(PERMUTE_FRONT_EDGES, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the front-face edges.
-pub const PERMUTE_FRONT_EDGES_270: [usize; NUM_EDGES] = compose(PERMUTE_FRONT_EDGES, 3);
+pub const PERMUTE_FRONT_EDGES_270: [u8; NUM_EDGES] = compose(PERMUTE_FRONT_EDGES, 3);
 
 /// The rotation matrix for a 180 degree turn of the back-face corners.
-pub const PERMUTE_BACK_CORNERS_180: [usize; NUM_CORNERS] = compose(PERMUTE_BACK_CORNERS, 2);
+pub const PERMUTE_BACK_CORNERS_180: [u8; NUM_CORNERS] = compose(PERMUTE_BACK_CORNERS, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the back-face corners.
-pub const PERMUTE_BACK_CORNERS_270: [usize; NUM_CORNERS] = compose(PERMUTE_BACK_CORNERS, 3);
+pub const PERMUTE_BACK_CORNERS_270: [u8; NUM_CORNERS] = compose(PERMUTE_BACK_CORNERS, 3);
 
 /// The rotation matrix for a 180 degree turn of the back-face edges.
-pub const PERMUTE_BACK_EDGES_180: [usize; NUM_EDGES] = compose(PERMUTE_BACK_EDGES, 2);
+pub const PERMUTE_BACK_EDGES_180: [u8; NUM_EDGES] = compose(PERMUTE_BACK_EDGES, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the back-face edges.
-pub const PERMUTE_BACK_EDGES_270: [usize; NUM_EDGES] = compose(PERMUTE_BACK_EDGES, 3);
+pub const PERMUTE_BACK_EDGES_270: [u8; NUM_EDGES] = compose(PERMUTE_BACK_EDGES, 3);
 
 /// The rotation matrix for a 180 degree turn of the left-face corners.
-pub const PERMUTE_LEFT_CORNERS_180: [usize; NUM_CORNERS] = compose(PERMUTE_LEFT_CORNERS, 2);
+pub const PERMUTE_LEFT_CORNERS_180: [u8; NUM_CORNERS] = compose(PERMUTE_LEFT_CORNERS, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the left-face corners.
-pub const PERMUTE_LEFT_CORNERS_270: [usize; NUM_CORNERS] = compose(PERMUTE_LEFT_CORNERS, 3);
+pub const PERMUTE_LEFT_CORNERS_270: [u8; NUM_CORNERS] = compose(PERMUTE_LEFT_CORNERS, 3);
 
 /// The rotation matrix for a 180 degree turn of the left-face edges.
-pub const PERMUTE_LEFT_EDGES_180: [usize; NUM_EDGES] = compose(PERMUTE_LEFT_EDGES, 2);
+pub const PERMUTE_LEFT_EDGES_180: [u8; NUM_EDGES] = compose(PERMUTE_LEFT_EDGES, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the left-face edges.
-pub const PERMUTE_LEFT_EDGES_270: [usize; NUM_EDGES] = compose(PERMUTE_LEFT_EDGES, 3);
+pub const PERMUTE_LEFT_EDGES_270: [u8; NUM_EDGES] = compose(PERMUTE_LEFT_EDGES, 3);
 
 /// The rotation matrix for a 180 degree turn of the right-face corners.
-pub const PERMUTE_RIGHT_CORNERS_180: [usize; NUM_CORNERS] = compose(PERMUTE_RIGHT_CORNERS, 2);
+pub const PERMUTE_RIGHT_CORNERS_180: [u8; NUM_CORNERS] = compose(PERMUTE_RIGHT_CORNERS, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the right-face corners.
-pub const PERMUTE_RIGHT_CORNERS_270: [usize; NUM_CORNERS] = compose(PERMUTE_RIGHT_CORNERS, 3);
+pub const PERMUTE_RIGHT_CORNERS_270: [u8; NUM_CORNERS] = compose(PERMUTE_RIGHT_CORNERS, 3);
 
 /// The rotation matrix for a 180 degree turn of the right-face edges.
-pub const PERMUTE_RIGHT_EDGES_180: [usize; NUM_EDGES] = compose(PERMUTE_RIGHT_EDGES, 2);
+pub const PERMUTE_RIGHT_EDGES_180: [u8; NUM_EDGES] = compose(PERMUTE_RIGHT_EDGES, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the right-face edges.
-pub const PERMUTE_RIGHT_EDGES_270: [usize; NUM_EDGES] = compose(PERMUTE_RIGHT_EDGES, 3);
+pub const PERMUTE_RIGHT_EDGES_270: [u8; NUM_EDGES] = compose(PERMUTE_RIGHT_EDGES, 3);
 
 /// The rotation matrix for a 180 degree turn of the up-face corners.
-pub const PERMUTE_UP_CORNERS_180: [usize; NUM_CORNERS] = compose(PERMUTE_UP_CORNERS, 2);
+pub const PERMUTE_UP_CORNERS_180: [u8; NUM_CORNERS] = compose(PERMUTE_UP_CORNERS, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the up-face corners.
-pub const PERMUTE_UP_CORNERS_270: [usize; NUM_CORNERS] = compose(PERMUTE_UP_CORNERS, 3);
+pub const PERMUTE_UP_CORNERS_270: [u8; NUM_CORNERS] = compose(PERMUTE_UP_CORNERS, 3);
 
 /// The rotation matrix for a 180 degree turn of the up-face edges.
-pub const PERMUTE_UP_EDGES_180: [usize; NUM_EDGES] = compose(PERMUTE_UP_EDGES, 2);
+pub const PERMUTE_UP_EDGES_180: [u8; NUM_EDGES] = compose(PERMUTE_UP_EDGES, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the up-face edges.
-pub const PERMUTE_UP_EDGES_270: [usize; NUM_EDGES] = compose(PERMUTE_UP_EDGES, 3);
+pub const PERMUTE_UP_EDGES_270: [u8; NUM_EDGES] = compose(PERMUTE_UP_EDGES, 3);
 
 /// The rotation matrix for a 180 degree turn of the down-face corners.
-pub const PERMUTE_DOWN_CORNERS_180: [usize; NUM_CORNERS] = compose(PERMUTE_DOWN_CORNERS, 2);
+pub const PERMUTE_DOWN_CORNERS_180: [u8; NUM_CORNERS] = compose(PERMUTE_DOWN_CORNERS, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the down-face corners.
-pub const PERMUTE_DOWN_CORNERS_270: [usize; NUM_CORNERS] = compose(PERMUTE_DOWN_CORNERS, 3);
+pub const PERMUTE_DOWN_CORNERS_270: [u8; NUM_CORNERS] = compose(PERMUTE_DOWN_CORNERS, 3);
 
 /// The rotation matrix for a 180 degree turn of the down-face edges.
-pub const PERMUTE_DOWN_EDGES_180: [usize; NUM_EDGES] = compose(PERMUTE_DOWN_EDGES, 2);
+pub const PERMUTE_DOWN_EDGES_180: [u8; NUM_EDGES] = compose(PERMUTE_DOWN_EDGES, 2);
 
 /// The rotation matrix for a 270 degree (counter-clockwise 90 degree) turn of the down-face edges.
-pub const PERMUTE_DOWN_EDGES_270: [usize; NUM_EDGES] = compose(PERMUTE_DOWN_EDGES, 3);
+pub const PERMUTE_DOWN_EDGES_270: [u8; NUM_EDGES] = compose(PERMUTE_DOWN_EDGES, 3);
